allow menu to store keybindings for quick menu navigation
